# WebApi/PortalTeleMedicina.WebAPI/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebAPI",
    "description": "Implementação pedida como teste para avaliação de conhecimento, para cargo de Desenvolver FullStack",
    "contact": {
      "name": "Marcelo Calixto",
      "url": "https://www.linkedin.com/in/marcelo-calixto-12182121/",
      "email": "marccalixto@hotmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Orders/{id}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/VendaVM"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Orders/PesquisarVendasPor": {
      "post": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PesquisarVendasPorDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ProdutoVM": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "ItemVendaVM": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "double",
          "type": "number"
        },
        "produto": {
          "$ref": "#/definitions/ProdutoVM"
        }
      }
    },
    "VendaVM": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        },
        "totalValue": {
          "format": "double",
          "type": "number"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ItemVendaVM"
          }
        },
        "usuarioId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "PesquisarVendasPorDTO": {
      "type": "object",
      "properties": {
        "dataInicial": {
          "format": "date-time",
          "type": "string"
        },
        "dataFinal": {
          "format": "date-time",
          "type": "string"
        },
        "valorInicial": {
          "format": "double",
          "type": "number"
        },
        "valorFinal": {
          "format": "double",
          "type": "number"
        },
        "usuarioId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}